Shave 61 bytes from the decompressor
This involved tons of little things, and I'm happy with the progress, but there are parts of it I would love to get down further.

Also I'm noticing that with both makeshift and witchcraft, it seems they don't work in the C64 debugger anymore (they _do_ work in VICE tho and I haven't tested on hw but I assume that would match VICE at least) and I'm really not sure why. The only thing I can think that really changed environment-wise is using different zp vars, but that should be safe as far as I know. Perhaps that's why makeshift doesn't actually make a clean exit tho; I might be using bad ones or something.